half broken url bar
The url bar currently stores the information in the url bar correctly for each tab and also correctly fetches the initial search query url only on the first tab, on any other tab you would need to switch to another tab and then go back to the previous tab to make it update the url. There is also an issue where after using the url bar to search for something and then navigating to another page, the title, url and favicon won't update.